20230715 查询站点分类修改

+++ zlt-commons/zlt-common-core/src/main/java/com/central/common/model/QuizSuborders.java
@@ -40,5 +40,5 @@ public class QuizSuborders extends SuperEntity {
     @ApiModelProperty(value = "赔率")
     private BigDecimal odds;
     @ApiModelProperty(value = "赔率2(三中二之中三，中特之中二）")
-    private Double odds2;
+    private BigDecimal odds2;
     }
